test: add test case for expected diagnostics
Switched from `jest` to `vscode-test` as the test runner to match the
recommended approach to running tests for VS Code [^1]. I chose to
ensure the new E2E test case was working over getting the old unit test
working to have more confidence that I wasn't breaking the extension
when adding new changes.

The new test case is largely based on the Ruff extension's test case
[^2].

The `tsconfig.json` config was updated to work with the types used in
newer Node.js APIs referenced in the added test case.

[^1]: https://code.visualstudio.com/api/working-with-extensions/testing-extension
[^2]: https://github.com/astral-sh/ruff-vscode/blob/660da85bdfee44a39fba7115eb35968362ec9b52/src/test/e2e.test.ts