feat: add settings for Fixinator config file & CommandBox args
This is the main contribution I want to make to the project, as the
additional configuration options make it easier to adopt the extension
at the company I work at.

I didn't add additional test cases for these features since it seemed
like overkill. Locally testing the changes seemed sufficient.

The logger was changed so that it's possible to filter the log output in
VSCode's "Output" panel using its built-in logic for log levels.

The code paths for the scans themselves were refactored to make it
easier to handle `cwd` related issues when a relative path for the
config file is specified and make them a little more flexible in the
future by passing the document URI, rather than just the file path.