feat: add config to payload for HTTPS scan
The HTTPS scan is a lot faster than using CommandBox, so I wanted to
make the HTTPS scan results match as closely with the CommandBox scan
results by passing the contents of the specified `.fixinator.json`
config file as a part of the payload, similar to how the CommandBox CLI
makes the request.

Additional changes were made to make debugging issues through the logs &
error handling clearer.